+++ wedding/src/main/java/project/wedding/card/constant/CardStatus.java
@@ -1,4 +1,4 @@
-package project.wedding.domain;
+package project.wedding.card.constant;
 
 public enum CardStatus {
     BACKLOG, PROGRESS, DONE
