+++ wedding/src/main/java/project/wedding/cardboard/repository/CardContainerRepository.java
@@ -1,0 +1,7 @@
+package project.wedding.cardboard.repository;
+
+import project.wedding.cardboard.domain.CardBoard;
+
+public interface CardContainerRepository {
+    CardBoard findByUserId(String userId);
+}
