--- wedding/src/main/java/project/wedding/domain/CardContainerRepository.java
@@ -1,5 +1,0 @@
-package project.wedding.domain;
-
-public interface CardContainerRepository {
-    CardBoard findByUserId(String userId);
-}
