+++ wedding/src/main/java/project/wedding/todo/domain/Todo.java
@@ -1,4 +1,4 @@
-package project.wedding.domain;
+package project.wedding.todo.domain;
 
 import lombok.Getter;
