+++ wedding/src/main/java/project/wedding/user/domain/User.java
@@ -1,0 +1,4 @@
+package project.wedding.user.domain;
+
+public class User {
+}
